Installer: On Vista, automatically close and restart Windows Explorer
Previously, the Restart Manager API available on Vista and above was only
used to identify the processes that lock a file which needs to be
installed. Now that API is also used to close and restart locking
processes in order to replace in-use files instead of using the rename
trick which requires the user to log off and on again for changes to take
effect.

Note that the FindProcessesUsingModules functions were changed to return a
DWORD which is the Restart Manager session handle on Vista and above. This
handle can be used in subsequent calls to the Restart Manager API.
Moreover, the Restart Manager session is no longer closed in the call to
FindProcessesUsingModules, so that needs to be done manually now.

For pre-Vista Windows versions nothing changed and the rename trick is
still used.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>