git-cheetah: Compile and use 64bit shell extension in the installer
On Windows 64bit the explorer is a 64bit process. Thus a 32bit dll can
not be loaded. To support the context menu plugin on 64bit windows we
compile a seperate 64bit dll of cheetah and register it in case the
installer is run on 64bit.

The fact that there are two architectures is now not visible in the
installer anymore since it automatically chooses the right dll.

Signed-off-by: Heiko Voigt <hvoigt@hvoigt.net>