Default the TERM environment variable suitable for use with less.
By default on windows the TERM environment variable is configured as
'winansi' in git unless previously set. This is not accepted by the default
pager (less) which produces an error message:
 "WARNING: terminal is not fully functional"
Under git-bash TERM is configured by the shell as cygwin which less accepts
and here we configure it as "msys" which also is acceptable to less.

Signed-off-by: Theo Niessink <niessink@martinic.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>