Fix inputrc for rxvt usage.
When using rxvt the backspace key deletes the whole line because for
rxvt and the bash prompt the backspace key produces different input.
This patch ammends the handling of backspace/control-backspace for rxvt
so that it only deletes one char at a time. Normal bash prompt input
is left as before.

Reported-by: Jon Seymour <jon.seymour@gmail.com>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>