Remove a bunch of extraneous ** from patterns in etc/gitignore
These were added in ef3e533f2e876425c17bb7d2d76d929703cc716b, but they
are hurting, not helping. Consider the following patterns:

foo/
**/foo/
foo/**

The first will match all directories named "foo" (and anything inside
them) no matter where they reside in the repository.

The second will match directories named "foo" (and anything inside
them), but only if they are below the .gitignore in the directory
structure.

The third will match anything inside all directories named "foo".

Because Git only tracks files, not directories, the first and third
patterns are essentially equivalent. Since the first is simpler, we
should prefer it.

The second pattern can be useful in some cases, but as far as I can tell
it is not actually desired in any of the cases where we are currently
using it.

So, I've replaced all instances of patterns 2 and 3 above with pattern
1.