Added a compiled wrapper to replace the git.cmd script.
This addresses github issue #36 which points out problems in handling
git commit notations of the form tag^{commit} and HEAD^. The windows
command shell uses caret as a quote both on the command line and in
batch scripts and this results in requiring double quoting. To be
able to handle both "tag^{commit}" and tag^^{commit} (either style
of command prompt quoting) the script needs replacing. This also makes
it simpler to call git from batch scripts as it will no longer need to be
prefixed with 'call'.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>