Include the CAcert root certificates in our CA certificate bundle.
This allows us to accept CAcert issued certificates as valid for SSL
connections. CAcert is a chain-of-trust certificate authority using a
model like PGP for authenticating trust and is used by some open source
sites as the CA issues free certificates provided there is sufficient
confirmation of identity using the web of trust.

The CAcert license requires mentioning in the documentation and it is
polite to list other packages that Git for Windows depends upon.

This closes github issue #37.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>