Make hacking on Git using an msysGit installer more firewall-friendly
Change the use of the git:// protocol for msysGit submodules to the
https:// protocol, so that developers behind firewalls who have
installed msysGit-netinstall or -fullinstall are not discouraged from
hacking on Git because 'git submodule update' fails to run.  Use the
https:// protocol instead of the http:// protocol so that the
developers are eventually able to push their changes back to origin.
The performance penalty for downgrading from the git:// protocol to the
https:// protocol is minimal, given that msysGit now uses a smart-HTTP-
aware Git.

Signed-off-by: Peter Klavins <klavins@netspace.net.au>