add script to build python
This is both nasty and incomplete, but at least for me it results
in a seemingly working python interpreter.

I don't know if or how complete the module-support is, but at least
all modules that seems to be built into the Windows EXE on MSVC are
present.

Perhaps someone else will enjoy this as well.

Signed-off-by: Erik Faye-Lund <kusmabite@gmail.com>