Improve Cygwin path-stripping code (fixes issue #67)
Using awk to strip any Cygwin path creates a null-byte at the end of the
string. This makes the last item in the path unresolvable and causes child
processes that are not dependent on msys-1.0.dll to inherit POSIX rather
than Windows paths. Fix this by not using awk at all.

Reported-by: John Stevenson <john-stevenson@blueyonder.co.uk>
Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>