Improve Cygwin path-stripping code
Using awk to strip the Cygwin paths creates a null-byte at the end of the string. This makes the last item in the path unresolvable and causes child processes that are not dependent on msys-1.0.dll to inherit POSIX rather than Windows paths. This routine avoids awk and includes code from nicerobot on StackOverflow: http://stackoverflow.com/a/370255/1127485

Signed-off-by: johnstevenson <john-stevenson@blueyonder.co.uk>