Replace 'devel' with 'master'
As discussed on the msysGit mailing list, we switched from the
two-integration branch workflow to a single one.

The original idea of having 'devel' and 'master' as two separate
integration branches was to let 'devel' be rebased all the time while
'master' *must* fast-forward always.

However, it turned out that such a maintainership style falls on its face
when there are more than a single maintainer, so we had to invent the
merging rebase strategy: it allows us to rebase but fast-forward at the
same time. It does so by faking a merge when going back to a clean
upstream revision and re-applying all the patches.

Since that distinction between 'devel' and 'master' is now obsolete, we
can just do away with 'devel'. It has served us well and deserves to be
retired now. May it rest in peace.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>