New msysGit / Git for Windows Logo
I'm a big fan of the Git for Windows project, so when I realized that the
project is still using Git's older logos, I decided to try my hand at
updating the logo.

As you can see, it's a bit more Windows friendly!

The original logo is licensed under the Creative Commons Attribution
License, so it allows modification.

I started with an SVG of the original logo and completely left the "git"
text alone while also being very careful to keep my modifications of the
orange diamond to the exact same dimensions. I even found the original
font used for the "git" text (Adelle bold) and added the "msys" title.

Hope you guys like it! If there's any interest in updating the project to
use this new logo, here are the SVGs.

[from: https://github.com/msysgit/git/issues/55]

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>