curl: switch back to openssl and avoid dynamic link to zlib
Erik noted that the new compilation options cause libcurl to be linked to
zlib1.dll which we do not include. Removing the ZLIB_PATH parameter
keeps it statically linked.

The winssl code can fail to operate behind a proxy. It tries to check for
certificate revocation, fails and then aborts the ssl negotiations.
Therefore reverting back to openssl for ssl support.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>