merging-rebase: be more suspicious about merging-rebase candidates
Sometimes people commit all kinds of stuff before performing a merging
rebase, and that unfortunately also entails merges that look like '-s
ours' type merges. Therefore, it is not good enough to look whether
the merge looks like that, but we also have to ask the user if that's
the start of the latest merging rebase.

Only when we find that the commit subject contains 'merging-rebase',
we trust that we have found the correct merge.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>