Installer: Detect TortoiseGitPlink from TortoiseGit
TortoiseGit comes with a Git optimized version of TortoisePlink called TortoiseGitPLink. A new name was chosen so that TortoisePLink provides the exact same results as the TortoiseSVN version.

Main change to TortoiseGitPlink is that the "-batch" parameter which causes problems in TortoisePlink is ignored (see http://code.google.com/p/tortoisegit/issues/detail?id=1617).

Signed-off-by: Sven Strickroth <email@cs-ware.de>