Do not set the HOME variable in the installer
We really should not need to set HOME globally anymore because we set it
in Git Bash in /etc/profile, and in the Git wrappers for cmd.exe.

Setting HOME caused problems, in particular when we hardcoded it to
the administrator's HOME if Git was installed using the administrator's
account.

This closes msysgit/git#119.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>