Installer: Do not mention Cygwin on the PATH page to avoid confusion
MSYS, being a Cygwin fork, also looks at the CYGWIN environment variable.
If a user has installed Cygwin in parallel to MSYS / Git for Windows,
CYGWIN might have been customized for Cygwin, but that customization also
affects MSYS and might break it as [1] shows. Do not explicitly mention
Cygwin as part of the installer's wizard page for adjusting the PATH
environment variable to not give the impression parallel use with Cygwin
was supported or encouraged. In fact, we highly recommend to not have
Cygwin installed at all when using MSYS / Git for Windows.

Also, take the opportunity to reword the other options on the same wizard
page for clarity.

[1] https://github.com/msysgit/git/issues/118

Signed-off-by: Thomas Braun <thomas.braun@byte-physics.de>
Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>