perl: add 0008-Fix-Term-ReadLine.patch
Add patch for Term::ReadLine.
https://github.com/msysgit/msysgit/issues/61#issuecomment-10695361

This patch has been directly applied to /lib/perl5/5.8.8
in db8d1bff1b121a08ca3782c5f52ff6a04ff98533
(see https://github.com/msysgit/msysgit/issues/152)
In this commit, the patch is put in /src/perl/patches to be applied
on top of the perl 5.8.8 source to allow a clean rebuild.

As the fix has been integrated in upstream in perl 5.18, it will not
have to be ported when upgrading perl.