Add the vimtutor shell script
This is the exact same vimtutor from the vim source repository:

  https://vim.googlecode.com/hg/src/vimtutor

It works as is in MSYS, although there are a couple of warnings:

  /bin/vimtutor: line 28: mktemp: command not found
  /bin/vimtutor: line 28: tempfile: command not found

Since these utilities are not present.

However, the script has a contingency for that scenario, and manages to
create a temporary copy of the tutor file anyway, so all is well.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>