Add a helper for ssh sessions outside Git Bash
Outside Git Bash, it can be hard to use the ssh-agent because it is not on
the PATH, and besides, there is no mechanism similar to .bashrc to start
the ssh-agent.

So let's add a helper script to /cmd/ that calls a shell script in /bin/
(that way, we get all the Bash niceties for free, without bleeding Bash
into the Git Cmd environment).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>