Use delayed expansion of environment variables
With the immediate expansion, paths containing the ampersand symbol (&)
will not work in this script. This is likely to occur when the user
profile is located in a folder that contains the symbol in its name.
The command processor misinterprets the symbol as to mean the union of
two separate commands.

To avoid the issue, "delayed expansion" is used and the environment
variables are demarcated using the exclamation mark (!) instead of
the percentage symbol (%).

Signed-off-by: ADTC <send2adtc@gmail.com>