Enable ambiguous completion on first Tab press
The show-all-if-ambiguous setting is now enabled in /etc/inputrc.

Previously the setting was set to off, accompanied by the comment 'Show
all instead of beeping first'. The documentation for this setting says:

  This alters the default behavior of the completion functions. If set
  to ‘on’, words which have more than one possible completion cause the
  matches to be listed immediately instead of ringing the bell. The
  default value is ‘off’.

Since the setting is off by default and setting it to off does the
opposite of what the comment says, it looks like the intent was to
enable the setting and that it's disabled by mistake. The commit
message in ed01c26 ('Kills the annoying bell') where this setting was
originally added also suggests that the setting was intended to be
enabled.