make vimdiff be able to be called by git mergetool
git mergetool does not work by default even though vimdiff (one of the
tools it tries to use by default) is installed. vimdiff can be called as
vim -d but not vimdiff from the shell.

Options:
1. call manually as vim -d
more things for user. I think I used this method like this:
	git mergetool --tool="vim -d"

2. link vimdiff to vim (a bit involved on Windows see [Symlinks])
3. make vimdiff.exe call vim.exe
seems slighty more complex but I don't know how all of msysgit
works

4. copy bin/vim to bin/vimdiff (requires adding -d options to exec
call)

I have attached a patch that does number four, since it is the least
complicated with the current features of the program, and makes seem
like the way it works on Linux (defaultable). the -d option on the exec
call is only neccesary because it calls the vim binary, not one called
vimdiff, if there was one called vimdiff it would not be needed, but
this method makes a script called vimdiff call vim -d.

Research:
While vimdiff is typically installed as a symbolic link, this is not
possible here as symbolic links are not supported on Windows (there is a
similar thing, called "reparse points", but it applies only to some
Windows versions and requires special privileges anyway).

[1] http://stackoverflow.com/questions/8876323/how-does-the-softlink-vimdiff-be-implemented
[2] http://www.quora.com/How-does-git-handle-symbolic-links-on-Windows
[3] http://stackoverflow.com/questions/5917249/git-symlinks-in-windows